Add a --stopOnKeyPress option to the cli
Note that this is a temporary workaround to allow  for runnig the brooklyn-whirr example.

This will be replaced by more powerful CLI control for running processes, to send shutdown and other commands to brooklyn.

Without using this flag you get a java.util.concurrent.RejectedExecutionException because the brooklyn and whirr shutdown hooks get executed in parallel.

This is how it looks like: {@linktourl https://gist.github.com/47066f72d6f6f79b953e}