AbstractController: use config/sensors instead of fields
Will make it easier to serialize/deserialize if using standard
mechanism of getAttribute rather than duplicating the values in
fields.

Also, "fixes" makeUrl which was called from checkEntity. It therefore
overwrote the value of SPECIFIED_URL attribute each time. And tests
only worked because of the call to makeUrl inside checkEntity, which
was very misleading.

I'm not sure about the old comment in preStart:
  previously we would attempt to infer values from a specified URL, but now we don't
  as that specified URL might be for another machine with port-forwarding
I've changed it so we really do believe the SPECIFIED_URL for reporting the
protocol and specified_url attributes.