Make embedded jars optional
Modifies code that relies on jar files as an embedded resource to
gracefully handle cases where the files are missing. Limited to tests,
which will cause TestNG to skip the test if the resource is missing.
Required for our source distribution, as Apache convention is to not
ship binaries in releases.