web app deploy - use temporary file for upload
Uploading the file directly to the webapps folder results in autodeploy triggering before the upload is complete. A problem arises when autodeploy triggers in the same second (before) the upload completes - it will fail due to incomplete file, but will not retry because the timestamp of the file won't change. Change deploy to upload to a temporary file and rename it to a .war only when done. This has a few benefits:
  * The application won't be undeployed at the start of the upload and be down for the duration of the upload
  * Once the server observers the updated file we can be sure that it read its full content
  * If the upload fails (timeout, connection reset, full disk...) the old app won't be undeployed.

The fix changes the semantics of the deploy method. Previously it always succeeded regardless of the result of the copy operation. Now it will fail if either the copy or the rename operations fail.